Updated to v1.3.0
* Added new managed (.NET C#) assembly compiler and C# program files
* Added option to choose between building native (C) or managed (.NET C#) builds
* Added "Current Directory" into the file "Drop Location" option
* Rewrote native build program code for fewer detections
* Obfuscated all commands and added command string morphing to avoid static string detection
* Changed Windows Defender exclusion commands to the new undetected form
* Added save and load functionality to the builder
* Added message box type selection to the "Fake Error" option
* Changed compilers to always compile with a manifest to reduce detections
* Restructured all project folders and files
* Updated compilers
